remote-tmux: remove the superseded pure-frame layout walk
RemoteTmuxMirrorGeometry.clientCells/frames and the mirror's
framesForRender had no callers outside their own tests: the native chrome
rewrite replaced the live pipeline with RemoteTmuxNativeLayoutMetrics for
the claim and bonsplit divider fractions for the render, but left the old
walk in the tree with doc comments still describing it as the sizing
authority. That cost us a real debugging detour — the comments point at
the wrong code when the mirror misrenders.

Keep the struct itself: it carries the measured render constants
(cell/padding/scale) the native metrics are built from, plus the claim
floors. Drop the dead functions, their frames value type, their test
suite, and the one feed-forward test that exercised the dead entry point,
and rewrite the mirror's header to describe the pipeline that actually
runs. The one still-live function the deleted suite covered
(patchingLeafRects) keeps its regression test.
